Reorganize modules in Cardano.Wallet.DB.*
Move

* manual database migrations
* `Checkpoints` type
* and the creation of a `Store` for checkpoints

into separate modules. The intention is that the store creation will be swapped out by a different implementation later on.
